requirements.txt
torch
easyocr
Pillow
pynput
